+++ src/Laravel/Eloquent/State/ItemProvider.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 
 namespace ApiPlatform\Laravel\Eloquent\State;
 
+use ApiPlatform\Laravel\Eloquent\Extension\QueryExtensionInterface;
+use ApiPlatform\Metadata\Exception\RuntimeException;
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\ProviderInterface;
 use Illuminate\Database\Eloquent\Model;
@@ -24,25 +26,37 @@ use Psr\Container\ContainerInterface;
 final class ItemProvider implements ProviderInterface
 {
     use LinksHandlerLocatorTrait;
+    use ModelClassTrait;
 
     /**
-     * @param LinksHandlerInterface<Model> $linksHandler
+     * @param LinksHandlerInterface<Model>      $linksHandler
+     * @param iterable<QueryExtensionInterface> $queryExtensions
      */
     public function __construct(
         private readonly LinksHandlerInterface $linksHandler,
         ?ContainerInterface $handleLinksLocator = null,
+        private iterable $queryExtensions = [],
     ) {
         $this->handleLinksLocator = $handleLinksLocator;
     }
 
     public function provide(Operation $operation, array $uriVariables = [], array $context = []): object|array|null
     {
-        $model = new ($operation->getClass())();
+        $resourceClass = $this->getModelClass($operation);
+        $model = new $resourceClass();
+
+        if (!$model instanceof Model) {
+            throw new RuntimeException(\sprintf('The class "%s" is not an Eloquent model.', $resourceClass));
+        }
 
         if ($handleLinks = $this->getLinksHandler($operation)) {
             $query = $handleLinks($model->query(), $uriVariables, ['operation' => $operation] + $context);
         } else {
             $query = $this->linksHandler->handleLinks($model->query(), $uriVariables, ['operation' => $operation] + $context);
+        }
+
+        foreach ($this->queryExtensions as $extension) {
+            $query = $extension->apply($query, $uriVariables, $operation, $context);
         }
 
         return $query->first();